marxel initial

+++ src/main/java/com/MP_121/repository/UsersRepository.java
@@ -4,7 +4,7 @@ import com.MP_121.model.UsersModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
-public interface UsersRepository extends JpaRepository<UsersModel, Integer>{
+public interface UsersRepository extends JpaRepository<UsersModel, Long>{
 
     Optional <UsersModel> findByEmailAndPassword(String email, String password);
 }
